guardiansoftherift: don't take weak cell at start

+++ guardiansoftherift/src/main/java/io.reisub.unethicalite.guardiansoftherift/tasks/TakeUnchargedCells.java
@@ -10,7 +10,7 @@ import net.unethicalite.api.commons.Time;
 import net.unethicalite.api.entities.TileObjects;
 import net.unethicalite.api.items.Inventory;
 
-public class TakeStartItems extends Task {
+public class TakeUnchargedCells extends Task {
 
   @Inject
   private GuardiansOfTheRift plugin;
@@ -19,7 +19,7 @@ public class TakeStartItems extends Task {
 
   @Override
   public String getStatus() {
-    return "Taking start items";
+    return "Taking uncharged cells";
   }
 
   @Override
@@ -27,28 +27,16 @@ public class TakeStartItems extends Task {
     return (plugin.getElapsedTicks() == -1
         || (plugin.getElapsedTicks() < 100 && Inventory.contains(ItemID.GUARDIAN_ESSENCE)))
         && GotrArea.getCurrent() == GotrArea.MAIN
-        && (!Inventory.contains("Weak cell")
-        || Inventory.getCount(true, "Uncharged cell") < 10);
+        && Inventory.getCount(true, "Uncharged cell") < 10;
   }
 
   @Override
   public void execute() {
-    if (!Inventory.contains("Weak cell")) {
-      if (Inventory.isFull()) {
-        Inventory.getFirst(ItemID.GUARDIAN_ESSENCE).drop();;
-      }
-
-      TileObjects.getNearest("Weak cells").interact("Take");
-      Time.sleepTicksUntil(() -> Inventory.contains("Weak cell"), 20);
+    if (Inventory.isFull() && !Inventory.contains("Uncharged cell")) {
+      Inventory.getFirst(ItemID.GUARDIAN_ESSENCE).drop();;
     }
 
-    if (Inventory.getCount(true, "Uncharged cell") < 10) {
-      if (Inventory.isFull() && !Inventory.contains("Uncharged cell")) {
-        Inventory.getFirst(ItemID.GUARDIAN_ESSENCE).drop();;
-      }
-
-      TileObjects.getNearest("Uncharged cells").interact("Take-10");
-      Time.sleepTicksUntil(() -> Inventory.getCount(true, "Uncharged cell") == 10, 20);
-    }
+    TileObjects.getNearest("Uncharged cells").interact("Take-10");
+    Time.sleepTicksUntil(() -> Inventory.getCount(true, "Uncharged cell") == 10, 20);
   }
 }
